pyproject.toml
[tool.poetry]
name = "logs-and-metas-for-stations"
version = "0.1.0"
description = ""
authors = ["codepanda <17920082@qq.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.2"
djangorestframework = "^3.11.0"
django-mptt = "^0.11.0"
PyYAML = "^5.3"
django-filter = "^2.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
pysocks = "^1.7.1"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
